Move Veryl module construction into frontend
