+++ crates/celox-slt/src/lib.rs
@@ -5,7 +5,7 @@
 
 use std::collections::BTreeSet;
 
-use celox_design::VarAtomBase;
+use celox_design::{ModuleId, VarAtomBase};
 use fxhash::{FxHashMap as HashMap, FxHashSet as HashSet};
 
 pub mod const_inline;
@@ -16,6 +16,7 @@ mod node_rules;
 mod path;
 pub mod range_store;
 pub mod scheduler;
+mod symbolic_verify;
 
 #[doc(hidden)]
 pub use lower::matches_slt_or_scan_group;
@@ -28,6 +29,7 @@ pub use node_facts::{SLTNodeFacts, SLTNodeFactsError};
 pub use path::{LogicPath, LogicPathId, LogicPathTarget};
 pub use range_store::{RangeStore, RangeStoreError};
 pub use scheduler::FfAccessSummary;
+pub use symbolic_verify::verify_symbolic_roots;
 
 /// Return the construction-time width cached when a node was interned.
 pub fn get_width<A: std::hash::Hash + Eq + Clone>(node: NodeId, arena: &SLTNodeArena<A>) -> usize {
@@ -62,6 +64,43 @@ pub struct CombObserver<A> {
     pub written_input_atoms: Vec<VarAtomBase<A>>,
     pub written_inputs: Vec<A>,
     pub captured_in_loop: bool,
+}
+
+#[derive(
+    Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
+)]
+pub enum GlueAddrBase<V> {
+    Parent(V),
+    Child(V),
+}
+
+impl<V: Copy> GlueAddrBase<V> {
+    pub fn var_id(&self) -> V {
+        match self {
+            GlueAddrBase::Parent(value) | GlueAddrBase::Child(value) => *value,
+        }
+    }
+}
+
+impl<V: std::fmt::Display> std::fmt::Display for GlueAddrBase<V> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            GlueAddrBase::Parent(value) => write!(f, "GlueAddr::Parent({value})"),
+            GlueAddrBase::Child(value) => write!(f, "GlueAddr::Child({value})"),
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+#[serde(bound(
+    serialize = "V: serde::Serialize",
+    deserialize = "V: serde::Deserialize<'de> + std::hash::Hash + Eq + Clone"
+))]
+pub struct GlueBlockBase<V: std::hash::Hash + Eq + Clone> {
+    pub module_id: ModuleId,
+    pub input_ports: Vec<(Vec<V>, LogicPath<GlueAddrBase<V>>)>,
+    pub output_ports: Vec<(Vec<V>, LogicPath<GlueAddrBase<V>>)>,
+    pub arena: SLTNodeArena<GlueAddrBase<V>>,
 }
 
 #[cfg(test)]
