+++ crates/celox/src/logic_tree.rs
@@ -1,15 +1,6 @@
 mod comb;
 pub use celox_slt::const_inline;
-pub use celox_slt::range_store;
 pub use comb::SLTLoopBound;
 pub use comb::SLTNode;
-pub(crate) use comb::SLTNodeArenaEditError;
-pub(crate) use comb::coerce_node_width;
-pub(crate) use comb::parse_comb_with_loop_recovery;
 pub use comb::{LogicPath, LogicPathTarget};
-pub use comb::{
-    NodeId, SLTForFoldGroupState, SLTNodeArena, SLTNodeFacts, SLTNodeFactsError, SymbolicStore,
-    eval_assignment_expression, eval_expression, get_width,
-};
-#[cfg(test)]
-pub(crate) use comb::{SLTForUpdate, SLTStepOp};
+pub use comb::{NodeId, SLTForFoldGroupState, SLTNodeArena, SLTNodeFactsError, get_width};
