--- crates/celox/src/parser/bitaccess.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::bitaccess::*;
