--- crates/celox/src/parser/bitslicer.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::bitslicer::*;
