--- crates/celox/src/parser/case.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::case::*;
