+++ docs/internals/compiler-crate-architecture.md
@@ -37,11 +37,13 @@ physical layout before the VM executes the bound bytecode. Native, Cranelift, an
 therefore share the same testbench binding path. Moving the remaining parser implementation,
 symbolic, optimizer, and runtime payload into the phase-specific target crates below remains.
 `celox-slt` now owns the frontend-independent bit-range store, symbolic-state contract, SLT node
-arena, iterative facts/verifier, path graph, fused comb/FF scheduler, and SLT-to-SIR lowerer. The FF
-lowering callback has an associated frontend-owned error type, so the scheduler no longer imports
-parser errors. `celox-frontend-veryl` owns lowering diagnostics, context-width and bit-select
-semantics, analyzer loop provenance, and the Veryl comb arena builder. Module/FF orchestration is
-the remaining frontend construction move.
+arena, iterative facts/verifier, symbolic-root verification, path graph, fused comb/FF scheduler,
+glue contract, and SLT-to-SIR lowerer. The FF lowering callback has an associated frontend-owned
+error type, so the scheduler no longer imports parser errors. `celox-frontend-veryl` owns lowering
+diagnostics, context-width and bit-select semantics, analyzer loop provenance, the Veryl comb arena
+builder, FF construction, module construction, and the resulting per-module symbolic artifact.
+Hierarchy discovery, flattening, relocation to design-owned identities, and final compiler-driver
+orchestration remain in the facade and are the remaining Milestone 4 ownership boundary.
 
 The baseline is the compiler pipeline on `perf/native-simulation-throughput` after PR #322. The
 split must preserve RTL semantics, generated-code quality, and the public `celox` API while making
